server/azure_functions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
# TODO: define the minium dependencies required
pyotp==2.4.0
PyJWT==1.7.1
facenet-pytorch==1.0.1
torch==1.3.1
torchvision==0.4.2
numpy==1.18.0
Pillow==7.2.0